requirements/local.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
crispy-forms-gds==0.2.1
cycler==0.10.0
distlib==0.3.1
Django==3.0.9
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-model-utils==4.1.1
filelock==3.0.12
idna==2.10
image==1.5.32
imutils==0.5.3
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
np==1.0.2
numpy==1.19.2
opencv-contrib-python==4.4.0.46
pathlib==1.0.1
Pillow==7.2.0
Pillow-PIL==0.1.dev0
psycopg2==2.8.6
pygame==2.0.0
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Unipath==1.1
urllib3==1.26.2
virtualenv==20.0.27
wrapt==1.12.1
